added directory watcher API..
fuck nio. Srsly. Who the fuck designed a reactive style base in this pattern? This is NOT timely. This is outdated. This is garbage. If you where to update this API, i will happily help design it. You do not have to change the base API, you can simply provide a wrapper that allows for reactive style. This is what is needed!