Srsly Oracle. Keep on reading.
I love Java. I really do. I think this is such beautiful language! So beautiful and so good for design. But you NEED to get some time of. For your next release, please, take your time and overwork your APIs, also, in specific your NIO API. This may be the best thing you could do. If you do not, you will end up with a historically grown language, that will be so cluttered, that no one will look at it in several years.